Core of mapping keyboard numberpad like iOS 12-key flick
This is currently missing a few flicks on the comma/numberpad_period
(need to send a shifted key - easy to do), and modifier support
(ten-ten, maru, small forms). The later can follow the approach used
in the New Sticknew mapping (send delete and the replacement presses,
based on a Karabiner Element's variable to retain state).