Start of a Hands Down Promethium on mapping on Japanese MacBook
Currently rule only active on JIS mode (KE virtual device driver mode),
but that means the punctuation needs to be adjusted (written assuming
ANSI, might work on ISO too for the keys considered).

Does not implement any combos, meaning no way to type tab, nor any of
the punctuation/symbols not on the number row. That includes [] and {},
as well as @, back-tick, +, :, * (assuming JIS mode).