Return left hand to Qwerty home, except bottom row
Trying to make the home vs bottom stagger as comfortable as possible,
and retain muscle memory for tab on this style keyboard.

We're effectively using "The Angle Mod" normally done on an ISO
keyboard moving ZXCVB left a key using the ISO key - but here with
a JIS MacBook keyboard that works by turning the left-shift into a
letter key.