Try left-hand phone-style (inverted) number-pad on nav-layer
This is not touch-type suitable as using the JIS 123 on the number
row (so left-hand must be moved up and left), but adding a second
layer via Karabiner-Elements seems fiddly? cf. holding thumb on
my split keyboard for a number-pad layer on the right.

This also means giving up on using the nav-layer left-hand home
keys (SNTH) as one-shot modifiers, which again looked tricky in
Karabiner-Elements. However with the Japanese MacBook keyboard
we have dedicated modifier keys so should work OK.